+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Adds getTag function to get the tag from a message.
 - Adds getSubject function to get the subject from a message.
 - Adds tests to ensure that the json payload has the proper values.
+- Adds test to ensure that required fields are in the payload.
 
 ### Updated
 - Refactor PostmarkTransportTest.